--- testtestest.py
@@ -1,4 +1,0 @@
-from main import Calculations as calcs
-
-while True:
-    print(str(calcs.speed()) + "\r")
